+++ lib/flutter_country_code.dart
@@ -1,4 +1,4 @@
 export 'src/country_code_picker/country.dart';
 export 'src/country_code_picker/country_code_picker.dart';
-export 'src/country_code_picker/functions.dart';
+export 'src/country_code_picker/country_code_helper.dart';
 export 'src/region_checker/phone_number_util.dart';