+++ lib/src/country_code_picker/country.dart
@@ -3,23 +3,23 @@ class Country {
   final String flag;
   final String countryCode;
   final String callingCode;
-  final Map nameTranslations;
+  final Map<String, String> nameTranslations;
 
-  const Country(
-    this.name,
-    this.flag,
-    this.countryCode,
-    this.callingCode,
-    this.nameTranslations,
-  );
+  const Country({
+    required this.name,
+    required this.flag,
+    required this.countryCode,
+    required this.callingCode,
+    required this.nameTranslations,
+  });
 
   factory Country.fromJson(Map<String, dynamic> json) {
     return new Country(
-      json['name'] as String,
-      json['flag'] as String,
-      json['country_code'] as String,
-      json['calling_code'] as String,
-      json['nameTranslations'] as Map,
+      name: json['name'] as String,
+      flag: json['flag'] as String,
+      countryCode: json['country_code'] as String,
+      callingCode: json['calling_code'] as String,
+      nameTranslations: json['nameTranslations'] as Map<String, String>,
     );
   }
 }